Allows multiple filters per program (#10)

+++ x/dynamic_module.cc
@@ -13,50 +13,12 @@ namespace Http {
 namespace DynamicModule {
 
 DynamicModule::~DynamicModule() {
+  __envoy_dynamic_module_v1_event_http_filter_destroy_(module_ctx_);
   ASSERT(handle_ != nullptr);
   dlclose(handle_);
-  std::filesystem::remove(file_path_for_cleanup_);
-}
-
-void DynamicModule::initModuleOnLocal(const ObjectFileLocationFilePath& location,
-                                      const std::string& config, const std::string& uuid) {
-  // To reload a module, we need to create a hard copy of the file and load that.
-  const std::filesystem::path file_path_absolute = std::filesystem::absolute(location.file_path);
-  file_path_for_cleanup_ = file_path_absolute.string() + "." + uuid;
-
-  std::filesystem::copy(file_path_absolute, file_path_for_cleanup_,
-                        std::filesystem::copy_options::recursive);
-
-  // Load the module with RTLD_LOCAL to avoid symbol conflicts with other modules.
-  handle_ = dlopen(file_path_for_cleanup_.c_str(), RTLD_LOCAL | RTLD_LAZY);
-  if (!handle_) {
-    throw EnvoyException(fmt::format("cannot load : {} error: {}", name_, dlerror()));
+  if (!file_path_for_cleanup_.empty()) {
+    std::filesystem::remove(file_path_for_cleanup_);
   }
-
-  // Initialize the module.
-  initModule(config);
-}
-
-void DynamicModule::initModuleOnInlineBytes(const ObjectFileLocationInlineBytes& location,
-                                            const std::string& config, const std::string& uuid) {
-
-  // To load a module from inline bytes, we need to create a temporary file and load that.
-  const std::filesystem::path tmpdir = std::filesystem::temp_directory_path();
-  file_path_for_cleanup_ = tmpdir / ("lib" + uuid + ".so");
-
-  // Copy the bytes to the file.
-  std::ofstream file(file_path_for_cleanup_, std::ios::binary);
-  file.write(location.inline_bytes.data(), location.inline_bytes.size());
-  file.close();
-
-  // Load the module with RTLD_LOCAL to avoid symbol conflicts with other modules.
-  handle_ = dlopen(file_path_for_cleanup_.c_str(), RTLD_LOCAL | RTLD_LAZY);
-  if (!handle_) {
-    throw EnvoyException(fmt::format("cannot load : {} error: {}", name_, dlerror()));
-  }
-
-  // Initialize the module.
-  initModule(config);
 }
 
 #define RESOLVE_SYMBOL_OR_THROW(symbol_type)                                                       \
@@ -66,19 +28,44 @@ void DynamicModule::initModuleOnInlineBytes(const ObjectFileLocationInlineBytes&
         fmt::format("cannot resolve symbol: {} error: {}", #symbol_type, dlerror()));              \
   }
 
-void DynamicModule::initModule(const std::string& config) {
-  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_module_init);
-  module_ctx_ =
-      __envoy_dynamic_module_v1_event_module_init_(const_cast<char*>(config.data()), config.size());
-  if (module_ctx_ == nullptr) {
-    throw EnvoyException(fmt::format("init function in {} failed", name_));
+void DynamicModule::initModuleOnLocal(const std::string_view location, const std::string& config) {
+  const std::filesystem::path file_path_absolute = std::filesystem::absolute(location);
+
+  // Test if the shared object is already openend.
+  handle_ = dlopen(file_path_absolute.c_str(), RTLD_NOLOAD | RTLD_LOCAL | RTLD_LAZY);
+  if (!handle_) {
+    // If the shared object is not opened, open it. Note that this runs on main thread.
+    handle_ = dlopen(file_path_absolute.c_str(), RTLD_LOCAL | RTLD_LAZY);
+    if (!handle_) {
+      throw EnvoyException(fmt::format("cannot load : {} error: {}", name_, dlerror()));
+    }
+    // Calls the program init function.
+    RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_program_init);
+    if (__envoy_dynamic_module_v1_event_program_init_() != 0) {
+      throw EnvoyException(fmt::format("program init in {} failed", name_));
+    }
   }
-  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_http_context_init);
-  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_http_request_headers);
-  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_http_request_body);
-  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_http_response_headers);
-  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_http_response_body);
-  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_http_destroy);
+
+  // Load the module with RTLD_LOCAL to avoid symbol conflicts with other modules.
+
+  // Initialize the module.
+  initHttpFilter(config);
+}
+
+void DynamicModule::initHttpFilter(const std::string& config) {
+  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_http_filter_init);
+  module_ctx_ = __envoy_dynamic_module_v1_event_http_filter_init_(const_cast<char*>(config.data()),
+                                                                  config.size());
+  if (module_ctx_ == nullptr) {
+    throw EnvoyException(fmt::format("http filter init in {} failed", name_));
+  }
+  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_http_filter_destroy);
+  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_http_filter_instance_init);
+  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_http_filter_instance_request_headers);
+  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_http_filter_instance_request_body);
+  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_http_filter_instance_response_headers);
+  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_http_filter_instance_response_body);
+  RESOLVE_SYMBOL_OR_THROW(__envoy_dynamic_module_v1_event_http_filter_instance_destroy);
 }
 
 #undef RESOLVE_SYMBOL_OR_THROW
